requirements.txt
mistralai
faiss-cpu
python-dotenv
rank-bm25
tqdm
